Introduce ability to join multiline comments for descriptions
Comments are treated fairly literally as they are found in protobuf files when the `protoc-gen-openapiv2` plugin is creating an Swagger doc.

If you define a multiline comment on an RPC or message and break it over several lines with newlines, then those newlines will be included in the Swagger doc within the `description` field. This behaviour forces you to have extremely long comment lines in order to maintain the proper paragraph structure of the comment content.

As a workaround to this, I have introduced a new option for the `protoc-gen-openapiv2` plugin named `join_multiline_comments` that, when enabled, will try to preserve the overall paragraph structure of a comment by replacing single newline instances with spaces, while continuing to preserve double newline instances in order to allow for paragraph breaks.